add support for stealing maps
This commit adds the possibility to create an ebpf program that will
try to use the maps from an already existing program.  When compiling
the ebpf program it is checked if the maps are present on the "other"
ebpf program, if yes a new map is NOT created but a reference to the
existing one is taken.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>